pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "token-miser"
version = "0.1.0"
description = "Benchmark Claude Code configuration packages"
requires-python = ">=3.13"
license = "Apache-2.0"
authors = [{ name = "Rubin Johnson" }]
keywords = ["claude", "claude-code", "testing", "devtools", "kanon", "loadout"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "loadout @ git+https://github.com/rubin-johnson/loadout.git@v0.1.0",
    "kanon-cli==1.2.0",
]

[project.scripts]
token-miser = "token_miser.__main__:main"

[dependency-groups]
dev = [
    "pytest>=9.0",
    "ruff>=0.11",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/token_miser"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["F", "E", "W", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

